A simple rearangement of some code, no functional changes
